TracAnnouncer: Part 2 of 7 - Restore db-related plugin history, refs #5774 and #9742.
Main feature of the code added here:
 * seamless schema upgrade by dedicated upgrade scripts - Trac core style

From the beginnung (see [3015]) this plugin's SQL driven schema check relied
on a SELECT raising an exeption for non-existing db table.
This has been discussed lately and found to be a flawed approach, that is even
known to break installations and interferes with db upgrades for Trac 0.13dev
and ultimately Trac 1.0 as well.

After introduction of the db schema version number for this plugin, table
existence testing is required one last time, current schema version is
registered, and only the registered schema version gets checked further on.

git-svn-id: http://trac-hacks.org/svn/announcerplugin/trunk@12297 7322e99d-02ea-0310-aa39-e9a107903beb