TracAnnouncer: Convert type to match db table definitions, refs #7975, #8056 and #10384.
These changes are based on work by olistudent, Stephen Anderson and rea.
I made sure, that we respect PEP8 as well, at least as far as Trac core does.

Thanks to all of you for testing, reports and suggestions towards a portable
fix, and - ultimately - patience to get it finally resolved.

git-svn-id: http://trac-hacks.org/svn/announcerplugin/trunk@12303 7322e99d-02ea-0310-aa39-e9a107903beb